desktophab-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
stripe==9.9.0
sqladmin==0.16.1
itsdangerous==2.2.0
email-validator==2.1.1
python-dotenv==1.0.1
slowapi==0.1.9
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
boto3==1.34.140